refactor(agent-server): move trajectory secret redaction out of file_router
Extract the redaction helpers into a dedicated security utility module
(_secret_redaction.py) so router files stay focused on HTTP handling,
mirroring the existing _secrets_exposure.py precedent.
