+++ openhands-agent-server/openhands/agent_server/file_router.py
@@ -25,6 +25,7 @@ from fastapi.responses import FileResponse
 from pydantic import BaseModel
 from starlette.background import BackgroundTask
 
+from openhands.agent_server._secret_redaction import redacted_file_bytes
 from openhands.agent_server.config import get_default_config
 from openhands.agent_server.models import Success
 from openhands.agent_server.server_details_router import update_last_execution_time
@@ -36,9 +37,7 @@ from openhands.sdk.git.utils import (
     run_git_command,
     validate_git_repository,
 )
-from openhands.sdk.llm.llm import LLM_SECRET_FIELDS
 from openhands.sdk.logger import get_logger
-from openhands.sdk.utils.pydantic_secrets import REDACTED_SECRET_VALUE
 
 
 class SubdirectoryEntry(BaseModel):
@@ -140,94 +139,12 @@ async def _download_file(path: str) -> FileResponse:
         )
 
 
-# Field names that carry credentials anywhere in a persisted conversation
-# payload (base_state.json, meta.json, events). Kept in sync with
-# ``LLM_SECRET_FIELDS`` and extended with the settings-level ``llm_api_key``
-# alias. A trajectory archive is meant to be shared for debugging, so these
-# must never leave the server as plaintext OR as recoverable ciphertext.
-_TRAJECTORY_SECRET_FIELDS: frozenset[str] = frozenset(
-    (*LLM_SECRET_FIELDS, "llm_api_key")
-)
-
-
-def _redact_secrets_in_obj(obj: object) -> bool:
-    """Recursively replace secret-bearing values with the redaction marker.
-
-    Returns ``True`` if anything was changed. Both plaintext and encrypted
-    (Fernet) secret material are masked, so the guarantee holds regardless of
-    whether ``OH_SECRET_KEY`` was configured when the conversation was
-    persisted.
-    """
-    changed = False
-    if isinstance(obj, dict):
-        for key, value in obj.items():
-            if (
-                key in _TRAJECTORY_SECRET_FIELDS
-                and isinstance(value, str)
-                and value != ""
-                and value != REDACTED_SECRET_VALUE
-            ):
-                obj[key] = REDACTED_SECRET_VALUE
-                changed = True
-            elif _redact_secrets_in_obj(value):
-                changed = True
-    elif isinstance(obj, list):
-        for item in obj:
-            if _redact_secrets_in_obj(item):
-                changed = True
-    return changed
-
-
-def _redacted_file_bytes(path: Path) -> bytes | None:
-    """Return redacted bytes for a JSON/JSONL file, or ``None`` if unchanged.
-
-    Persisted conversation files are JSON (``*.json``) or newline-delimited
-    JSON (``*.jsonl``/the per-event log files). Anything that does not parse as
-    JSON is left untouched (returns ``None``) so the archive is byte-identical
-    for non-secret content.
-    """
-    suffix = path.suffix.lower()
-    if suffix not in (".json", ".jsonl"):
-        return None
-    try:
-        raw = path.read_text(encoding="utf-8")
-    except (OSError, UnicodeDecodeError):
-        return None
-
-    if suffix == ".jsonl":
-        changed = False
-        out_lines: list[str] = []
-        for line in raw.splitlines(keepends=True):
-            stripped = line.strip()
-            if not stripped:
-                out_lines.append(line)
-                continue
-            try:
-                obj = json.loads(stripped)
-            except json.JSONDecodeError:
-                out_lines.append(line)
-                continue
-            if _redact_secrets_in_obj(obj):
-                changed = True
-            newline = "\n" if line.endswith("\n") else ""
-            out_lines.append(json.dumps(obj) + newline)
-        return "".join(out_lines).encode("utf-8") if changed else None
-
-    try:
-        obj = json.loads(raw)
-    except json.JSONDecodeError:
-        return None
-    if not _redact_secrets_in_obj(obj):
-        return None
-    return json.dumps(obj).encode("utf-8")
-
-
 def _create_zip_from_directory(source_dir: Path, output_path: Path) -> None:
     """Create a zip archive for source_dir using only Python stdlib APIs.
 
     Secret-bearing fields (LLM/AWS credentials) in the persisted JSON payloads
     are redacted on the way into the archive so a downloaded trajectory never
-    leaks API keys — see ``_redacted_file_bytes``.
+    leaks API keys — see ``redacted_file_bytes``.
     """
     try:
         with zipfile.ZipFile(output_path, "w", zipfile.ZIP_DEFLATED) as archive:
@@ -235,7 +152,7 @@ def _create_zip_from_directory(source_dir: Path, output_path: Path) -> None:
             for path in sorted(source_dir.rglob("*")):
                 arcname = str(path.relative_to(source_dir.parent))
                 if path.is_file():
-                    redacted = _redacted_file_bytes(path)
+                    redacted = redacted_file_bytes(path)
                     if redacted is not None:
                         archive.writestr(arcname, redacted)
                         continue
